implemented JWT token verification for user routes

+++ app/routers/user.py
@@ -2,7 +2,7 @@ from re import M
 from fastapi import Response, status, HTTPException, Depends
 from fastapi.routing import APIRouter
 from ..database import get_db
-from .. import schemas, utils, models
+from .. import schemas, utils, models, oauth2
 from sqlalchemy.orm import Session
 
 router = APIRouter(
@@ -29,20 +29,26 @@ def create_user(user: schemas.UserCreate, db: Session = Depends(get_db)):
 
 
 @router.get("/{id}", status_code=status.HTTP_200_OK, response_model=schemas.UserResponse)
-def get_user_by_id(id: int, db: Session = Depends(get_db)):
+def get_user_by_id(id: int, db: Session = Depends(get_db), 
+current_user: int = Depends(oauth2.get_current_user)):
     user = db.query(models.User).filter(models.User.id == id).first()
     if user is None:
         raise HTTPException(status.HTTP_404_NOT_FOUND, f"User with id: {id} does not exist.")
+    if id != current_user.id:
+        raise HTTPException(status.HTTP_403_FORBIDDEN, detail=f"Not authorized to perform requested action")
 
     return user
 
 
 @router.put("/{id}", response_model=schemas.UserResponse)
-def update_user(id: int, user: schemas.UserCreate, db: Session = Depends(get_db)):
+def update_user(id: int, user: schemas.UserCreate, db: Session = Depends(get_db), 
+current_user: int = Depends(oauth2.get_current_user)):
     target_user_query = db.query(models.User).filter(models.User.id == id)
     target_user = target_user_query.first()
     if target_user is None:
         raise HTTPException(status.HTTP_404_NOT_FOUND, f"User with id: {id} does not exist")
+    if id != current_user.id:
+        raise HTTPException(status.HTTP_403_FORBIDDEN, detail=f"Not authorized to perform requested action")
     user.password = utils.hash(user.password)
     target_user_query.update(user.dict(), synchronize_session=False)
     db.commit()
@@ -52,11 +58,14 @@ def update_user(id: int, user: schemas.UserCreate, db: Session = Depends(get_db)
 
 
 @router.delete("/{id}")
-def deactivate_user(id: int, db: Session = Depends(get_db)):
+def delete_user(id: int, db: Session = Depends(get_db), 
+current_user: int = Depends(oauth2.get_current_user)):
     target_user_query = db.query(models.User).filter(models.User.id == id)
     user = target_user_query.first()
     if user is None:
         raise HTTPException(status.HTTP_404_NOT_FOUND, f"User with id: {id} does not exist.")
+    if id != current_user.id:
+        raise HTTPException(status.HTTP_403_FORBIDDEN, detail=f"Not authorized to perform requested action")
     target_user_query.delete(synchronize_session=False)
     db.commit()
